♻️ Refactor main to renderIndex
