+++ index.html
@@ -13,29 +13,6 @@
       </a>
     </header>
 
-    <!-- Options -->
-    <main id="main-container" class="pt-12 mx-4 mb-8 md:max-w-2xl md:mx-auto md:my-0">
-      <section id="options"></section>
-      <button id="submit-button" class="mt-7 primary-button" type="button">밥 먹을 준비 완료!</button>
-    </main>
-
-    <!-- Loading -->
-    <aside class="fixed z-50 flex-col hidden w-48 h-48 p-5 bg-white rounded flex-center pos-center">
-      <div class="relative w-10 h-10 m-auto animate-spin">
-        <i class="absolute top-0 left-0 inline-block w-4 h-4 rounded-full opacity-25 bg-blue-primary"></i>
-        <i class="absolute bottom-0 left-0 inline-block w-4 h-4 rounded-full opacity-50 bg-blue-primary"></i>
-        <i class="absolute top-0 right-0 inline-block w-4 h-4 rounded-full opacity-75 bg-blue-primary"></i>
-        <i class="absolute bottom-0 right-0 inline-block w-4 h-4 rounded-full bg-blue-primary"></i>
-      </div>
-      <p>메뉴 탐색 중...</p>
-    </aside>
-    <aside class="fixed z-40 hidden bg-black pos-full opacity-20"></aside>
-
-    <!-- Result -->
-    <main id="result-container" class="hidden pt-12 mx-4 mb-8 md:max-w-2xl md:mx-auto md:my-0">
-      <button id="reset-button" class="mt-7 primary-button" type="button">다시 찾아보기</button>
-    </main>
-
     <div id="app"></div>
     <script type="module" src="./src/main.js"></script>
   </body>
